feat(vpc): update the vpc suppression and add flow logs
